requirements.txt
beautifulsoup4==4.8.1
lxml==4.2.1
python-dateutil==2.6.1
Flask==1.1.1